# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BigFin API
  version: 0.1.0
  description: |
    Multi-tenant loan administration API powering GreenBucks and other operator apps.

    ## Key Concepts
    - **Tenant**: Operator app (e.g., GreenBucks) with isolated data
    - **EndCustomer**: Borrower or lender on the platform
    - **LoanOffer**: Proposed loan terms from lender to borrower
    - **LoanContract**: Accepted offer with repayment schedule
    - **PaymentIntent**: Idempotent payment request (disbursement or repayment)

    ## Idempotency
    All POST endpoints that create resources or initiate payments accept an `Idempotency-Key` header.
    Requests with the same key within 24 hours return the original response.

    ## Error Codes
    See the Error Taxonomy section at the bottom of this spec.
  contact:
    name: BigFin API Support
  license:
    name: Proprietary
servers:
  - url: https://api.bigfin.io/v1
    description: Production
  - url: https://sandbox.api.bigfin.io/v1
    description: Sandbox
security:
  - BearerAuth: []
...
paths:
...
  /customers:
    post:
      tags: [Customers]
      summary: Create end customer
      description: Register a new borrower or lender
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [borrower, lender, both]
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
  /customers/{customer_id}:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      tags: [Customers]
      summary: Get customer details
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Customers]
      summary: Update customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{customer_id}/kyc:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      tags: [Customers]
      summary: Get KYC status
      responses:
        '200':
          description: KYC status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycStatus'

    post:
      tags: [Customers]
      summary: Submit KYC verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycSubmission'
      responses:
        '202':
          description: KYC submission accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycStatus'

  # ============ FUNDING INSTRUMENTS ============
...
  /loan-contracts/{contract_id}/documents:
    parameters:
      - $ref: '#/components/parameters/ContractId'
    post:
      tags: [Documents]
      summary: Upload document
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, document_type]
              properties:
                file:
                  type: string
                  format: binary
                document_type:
                  type: string
                  enum: [receipt, statement, contract, amendment, other]
                description:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'

    get:
      tags: [Documents]
      summary: List loan documents
      parameters:
        - name: document_type
          in: query
          schema:
            type: string
            enum: [receipt, statement, contract, amendment, other]
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
  /documents/{document_id}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: Get document metadata
      responses:
        '200':
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /documents/{document_id}/download:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: Download document
      responses:
        '200':
          description: Document file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/*:
              schema:
                type: string
                format: binary

  # ============ WEBHOOKS ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Unique key for idempotent requests. Same key within 24h returns cached response.
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CustomerId:
      name: customer_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ContractId:
      name: contract_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        external_id:
          type: string
          description: Client-provided ID
        role:
          type: string
          enum: [borrower, lender, both]
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        business_name:
          type: string
        kyc_level:
          type: string
          enum: [none, basic, enhanced]
        risk_tier:
          type: string
          enum: [new, established, trusted]
        risk_flags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCustomerRequest:
      type: object
      required: [role, email, first_name, last_name]
      properties:
        external_id:
          type: string
        role:
          type: string
          enum: [borrower, lender, both]
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        business_name:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
    UpdateCustomerRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
    CustomerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        next_cursor:
          type: string
        has_more:
          type: boolean
    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          default: US
    KycStatus:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        level:
          type: string
          enum: [none, basic, enhanced]
        status:
          type: string
          enum: [not_started, pending, approved, rejected, expired]
        submitted_at:
          type: string
          format: date-time
        verified_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        rejection_reason:
          type: string
    KycSubmission:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [basic, enhanced]
        ssn_last4:
          type: string
          pattern: '^\d{4}$'
        date_of_birth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
        document_ids:
          type: array
          items:
            type: string
            format: uuid

    # ===== Funding Instruments =====
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        document_type:
          type: string
          enum: [receipt, statement, contract, amendment, other]
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
        sha256:
          type: string
        description:
          type: string
        uploaded_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    DocumentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        next_cursor:
          type: string
        has_more:
          type: boolean

    # ===== Webhooks =====
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          description: Additional context
        request_id:
          type: string
          description: Request ID for support
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "invalid_request"
            message: "Principal must be between $100 and $50,000"
            request_id: "req_abc123"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "not_found"
            message: "Loan contract not found"
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "invalid_state"
            message: "Offer has already been accepted"

# ===== Error Taxonomy =====
#
# Error codes follow the pattern: {category}_{specific}
#
# Authentication/Authorization:
#   - unauthorized: Missing or invalid token
#   - forbidden: Valid token but insufficient permissions
#   - step_up_required: Sensitive action requires re-authentication
#
# Validation:
#   - invalid_request: Malformed request body
#   - invalid_parameter: Specific parameter validation failed
#   - terms_out_of_policy: Loan terms violate product policy
#
# State:
#   - invalid_state: Action not allowed in current state
#   - already_exists: Duplicate resource (use idempotency key)
#   - not_found: Resource doesn't exist
#
# Payment:
#   - insufficient_funds: Not enough balance
#   - instrument_invalid: Funding instrument can't be used
#   - payment_failed: Payment provider rejected
#   - payment_returned: Payment was returned/reversed
#
# Limits:
#   - limit_exceeded: Transaction or balance limit hit
#   - rate_limited: Too many requests
#
# System:
#   - internal_error: Unexpected server error
#   - service_unavailable: Temporary outage
#   - provider_error: Payment provider issue